refactor: Переименовал файлы некомпонент в kebab-case.
