+++ NFT-Marketplace/src/data/how-it-works.js
